binding/c: autogen request api
