--- src/mpi/request/test.c
@@ -1,126 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-
-/* -- Begin Profiling Symbol Block for routine MPI_Test */
-#if defined(HAVE_PRAGMA_WEAK)
-#pragma weak MPI_Test = PMPI_Test
-#elif defined(HAVE_PRAGMA_HP_SEC_DEF)
-#pragma _HP_SECONDARY_DEF PMPI_Test  MPI_Test
-#elif defined(HAVE_PRAGMA_CRI_DUP)
-#pragma _CRI duplicate MPI_Test as PMPI_Test
-#elif defined(HAVE_WEAK_ATTRIBUTE)
-int MPI_Test(MPI_Request * request, int *flag, MPI_Status * status)
-    __attribute__ ((weak, alias("PMPI_Test")));
-#endif
-/* -- End Profiling Symbol Block */
-
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-#undef MPI_Test
-#define MPI_Test PMPI_Test
-#endif
-
-/*@
-    MPI_Test  - Tests for the completion of a request
-
-Input Parameters:
-. request - MPI request (handle)
-
-Output Parameters:
-+ flag - true if operation completed (logical)
-- status - status object (Status).  May be 'MPI_STATUS_IGNORE'.
-
-.N ThreadSafe
-
-.N waitstatus
-
-.N Fortran
-
-.N FortranStatus
-
-.N Errors
-.N MPI_SUCCESS
-.N MPI_ERR_REQUEST
-.N MPI_ERR_ARG
-@*/
-int MPI_Test(MPI_Request * request, int *flag, MPI_Status * status)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Request *request_ptr = NULL;
-    MPIR_Comm *comm_ptr = NULL;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TEST);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-
-    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    MPIR_FUNC_TERSE_REQUEST_ENTER(MPID_STATE_MPI_TEST);
-
-    /* Validate parameters, especially handles needing to be converted */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            MPIR_ERRTEST_ARGNULL(request, "request", mpi_errno);
-            MPIR_ERRTEST_REQUEST_OR_NULL(*request, mpi_errno);
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif /* HAVE_ERROR_CHECKING */
-
-    MPIR_Request_get_ptr(*request, request_ptr);
-
-    /* Validate parameters and objects (post conversion) */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            if (*request != MPI_REQUEST_NULL) {
-                /* Validate request_ptr */
-                MPIR_Request_valid_ptr(request_ptr, mpi_errno);
-                if (mpi_errno)
-                    goto fn_fail;
-            }
-
-            MPIR_ERRTEST_ARGNULL(flag, "flag", mpi_errno);
-            /* NOTE: MPI_STATUS_IGNORE != NULL */
-            MPIR_ERRTEST_ARGNULL(status, "status", mpi_errno);
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif /* HAVE_ERROR_CHECKING */
-
-    /* ... body of routine ...  */
-
-    /* MPIR_Test may free request_ptr, so make a copy of comm_ptr before calling MPIR_Test.
-     * comm_ptr will be used later at fn_fail when MPIR_Test returns an error. */
-    if (request_ptr)
-        comm_ptr = request_ptr->comm;
-    mpi_errno = MPIR_Test(request, flag, status);
-    if (mpi_errno)
-        goto fn_fail;
-
-    /* ... end of body of routine ... */
-
-  fn_exit:
-    MPIR_FUNC_TERSE_REQUEST_EXIT(MPID_STATE_MPI_TEST);
-    MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        mpi_errno =
-            MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__, MPI_ERR_OTHER,
-                                 "**mpi_test", "**mpi_test %p %p %p", request, flag, status);
-    }
-#endif
-    mpi_errno = MPIR_Err_return_comm(comm_ptr, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
