--- src/mpi/request/testall.c
@@ -1,131 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-
-/* -- Begin Profiling Symbol Block for routine MPI_Testall */
-#if defined(HAVE_PRAGMA_WEAK)
-#pragma weak MPI_Testall = PMPI_Testall
-#elif defined(HAVE_PRAGMA_HP_SEC_DEF)
-#pragma _HP_SECONDARY_DEF PMPI_Testall  MPI_Testall
-#elif defined(HAVE_PRAGMA_CRI_DUP)
-#pragma _CRI duplicate MPI_Testall as PMPI_Testall
-#elif defined(HAVE_WEAK_ATTRIBUTE)
-int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
-                MPI_Status array_of_statuses[]) __attribute__ ((weak, alias("PMPI_Testall")));
-#endif
-/* -- End Profiling Symbol Block */
-
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-#undef MPI_Testall
-#define MPI_Testall PMPI_Testall
-#endif
-
-/*@
-    MPI_Testall - Tests for the completion of all previously initiated
-    requests
-
-Input Parameters:
-+ count - lists length (integer)
-- array_of_requests - array of requests (array of handles)
-
-Output Parameters:
-+ flag - True if all requests have completed; false otherwise (logical)
-- array_of_statuses - array of status objects (array of Status).  May be
- 'MPI_STATUSES_IGNORE'.
-
-Notes:
-  'flag' is true only if all requests have completed.  Otherwise, flag is
-  false and neither the 'array_of_requests' nor the 'array_of_statuses' is
-  modified.
-
-If one or more of the requests completes with an error, 'MPI_ERR_IN_STATUS' is
-returned.  An error value will be present is elements of 'array_of_status'
-associated with the requests.  Likewise, the 'MPI_ERROR' field in the status
-elements associated with requests that have successfully completed will be
-'MPI_SUCCESS'.  Finally, those requests that have not completed will have a
-value of 'MPI_ERR_PENDING'.
-
-While it is possible to list a request handle more than once in the
-'array_of_requests', such an action is considered erroneous and may cause the
-program to unexecpectedly terminate or produce incorrect results.
-
-.N ThreadSafe
-
-.N waitstatus
-
-.N Fortran
-
-.N Errors
-.N MPI_SUCCESS
-.N MPI_ERR_IN_STATUS
-.N MPI_ERR_REQUEST
-.N MPI_ERR_ARG
-@*/
-int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
-                MPI_Status array_of_statuses[])
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TESTALL);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-
-    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    MPIR_FUNC_TERSE_REQUEST_ENTER(MPID_STATE_MPI_TESTALL);
-
-    /* Check the arguments */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            int i = 0;
-
-            MPIR_ERRTEST_COUNT(count, mpi_errno);
-
-            if (count != 0) {
-                MPIR_ERRTEST_ARGNULL(array_of_requests, "array_of_requests", mpi_errno);
-                /* NOTE: MPI_STATUSES_IGNORE != NULL */
-                MPIR_ERRTEST_ARGNULL(array_of_statuses, "array_of_statuses", mpi_errno);
-            }
-            MPIR_ERRTEST_ARGNULL(flag, "flag", mpi_errno);
-
-            for (i = 0; i < count; i++) {
-                MPIR_ERRTEST_ARRAYREQUEST_OR_NULL(array_of_requests[i], i, mpi_errno);
-            }
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif /* HAVE_ERROR_CHECKING */
-
-    /* ... body of routine ...  */
-
-    mpi_errno = MPIR_Testall(count, array_of_requests, flag, array_of_statuses);
-    if (mpi_errno != MPI_SUCCESS)
-        goto fn_fail;
-
-    /* ... end of body of routine ... */
-
-  fn_exit:
-
-    MPIR_FUNC_TERSE_REQUEST_EXIT(MPID_STATE_MPI_TESTALL);
-    MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        mpi_errno =
-            MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE, __func__, __LINE__, MPI_ERR_OTHER,
-                                 "**mpi_testall", "**mpi_testall %d %p %p %p", count,
-                                 array_of_requests, flag, array_of_statuses);
-    }
-#endif
-    mpi_errno = MPIR_Err_return_comm(NULL, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
