--- src/mpi/request/waitall.c
@@ -1,122 +1,0 @@
-/*
- * Copyright (C) by Argonne National Laboratory
- *     See COPYRIGHT in top-level directory
- */
-
-#include "mpiimpl.h"
-
-/* -- Begin Profiling Symbol Block for routine MPI_Waitall */
-#if defined(HAVE_PRAGMA_WEAK)
-#pragma weak MPI_Waitall = PMPI_Waitall
-#elif defined(HAVE_PRAGMA_HP_SEC_DEF)
-#pragma _HP_SECONDARY_DEF PMPI_Waitall  MPI_Waitall
-#elif defined(HAVE_PRAGMA_CRI_DUP)
-#pragma _CRI duplicate MPI_Waitall as PMPI_Waitall
-#elif defined(HAVE_WEAK_ATTRIBUTE)
-int MPI_Waitall(int count, MPI_Request array_of_requests[], MPI_Status array_of_statuses[])
-    __attribute__ ((weak, alias("PMPI_Waitall")));
-#endif
-/* -- End Profiling Symbol Block */
-
-/* Define MPICH_MPI_FROM_PMPI if weak symbols are not supported to build
-   the MPI routines */
-#ifndef MPICH_MPI_FROM_PMPI
-#undef MPI_Waitall
-#define MPI_Waitall PMPI_Waitall
-#endif
-
-/*@
-    MPI_Waitall - Waits for all given MPI Requests to complete
-
-Input Parameters:
-+ count - list length (integer)
-- array_of_requests - array of request handles (array of handles)
-
-Output Parameters:
-. array_of_statuses - array of status objects (array of Statuses).  May be
-  'MPI_STATUSES_IGNORE'.
-
-Notes:
-
-If one or more of the requests completes with an error, 'MPI_ERR_IN_STATUS' is
-returned.  An error value will be present is elements of 'array_of_status'
-associated with the requests.  Likewise, the 'MPI_ERROR' field in the status
-elements associated with requests that have successfully completed will be
-'MPI_SUCCESS'.  Finally, those requests that have not completed will have a
-value of 'MPI_ERR_PENDING'.
-
-While it is possible to list a request handle more than once in the
-array_of_requests, such an action is considered erroneous and may cause the
-program to unexecpectedly terminate or produce incorrect results.
-
-.N waitstatus
-
-.N ThreadSafe
-
-.N Fortran
-
-.N Errors
-.N MPI_SUCCESS
-.N MPI_ERR_REQUEST
-.N MPI_ERR_ARG
-.N MPI_ERR_IN_STATUS
-@*/
-int MPI_Waitall(int count, MPI_Request array_of_requests[], MPI_Status array_of_statuses[])
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_WAITALL);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-
-    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    MPIR_FUNC_TERSE_REQUEST_ENTER(MPID_STATE_MPI_WAITALL);
-
-    /* Check the arguments */
-#ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-            int i;
-            MPIR_ERRTEST_COUNT(count, mpi_errno);
-
-            if (count != 0) {
-                MPIR_ERRTEST_ARGNULL(array_of_requests, "array_of_requests", mpi_errno);
-                /* NOTE: MPI_STATUSES_IGNORE != NULL */
-
-                MPIR_ERRTEST_ARGNULL(array_of_statuses, "array_of_statuses", mpi_errno);
-            }
-
-            for (i = 0; i < count; i++) {
-                MPIR_ERRTEST_ARRAYREQUEST_OR_NULL(array_of_requests[i], i, mpi_errno);
-            }
-        }
-        MPID_END_ERROR_CHECKS;
-    }
-#endif /* HAVE_ERROR_CHECKING */
-
-    /* ... body of routine ...  */
-
-    mpi_errno = MPIR_Waitall(count, array_of_requests, array_of_statuses);
-    if (mpi_errno)
-        goto fn_fail;
-
-    /* ... end of body of routine ... */
-
-  fn_exit:
-    MPIR_FUNC_TERSE_REQUEST_EXIT(MPID_STATE_MPI_WAITALL);
-    MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    return mpi_errno;
-
-  fn_fail:
-    /* --BEGIN ERROR HANDLING-- */
-#ifdef HAVE_ERROR_CHECKING
-    mpi_errno = MPIR_Err_create_code(mpi_errno, MPIR_ERR_RECOVERABLE,
-                                     __func__, __LINE__, MPI_ERR_OTHER,
-                                     "**mpi_waitall",
-                                     "**mpi_waitall %d %p %p",
-                                     count, array_of_requests, array_of_statuses);
-#endif
-    mpi_errno = MPIR_Err_return_comm(NULL, __func__, mpi_errno);
-    goto fn_exit;
-    /* --END ERROR HANDLING-- */
-}
